Create student.js
You are required to sort an array of objects, where each student is represented as an object 

 Each object has 4 keys, i.e, name, chemistry marks, biology marks, and date of birth. You should sort the array of objects according to the following rules: 

1. The first priority is the total marks.
2. In case the students have the same total marks, the next priority should be the number of marks in biology.
3. If students have the same marks in biology, and chemistry and have the same date of birth, then any order between them is acceptable. 
Note: Date of birth is given in the DD-MM-YYYY format. Test your application with the data from the above link. There are 2 test cases provided for which you should try them out. 



Check sample test cases in the following link, where you'll find it inside of students.txt

https://drive.google.com/drive/folders/1s9lq68Ta89DZ3FYpCY_FyINLGCB-VQ0x?usp=sharing